:construction: Update calc
完成计算器计算部分基本功能，剩余部分运算符未处理

+++ Codes/VB/05_calc/calculate/dll.h
@@ -7,6 +7,6 @@
 #define DLLIMPORT __declspec(dllimport)
 #endif
 
-extern "C" DLLIMPORT __stdcall float compute(char *expression);
+extern "C" DLLIMPORT __stdcall double compute(char *expression);
 
 #endif
